+++ src/components/Sidebar.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { useSelector, useDispatch } from 'react-redux';
-import { Segment, Accordion, Form, Grid } from 'semantic-ui-react';
+import { useDispatch } from 'react-redux';
+import { Segment, Accordion, Form } from 'semantic-ui-react';
 import { defineMessages, useIntl, FormattedMessage } from 'react-intl';
 
 import Icon from '@plone/volto/components/theme/Icon/Icon';
@@ -11,16 +11,11 @@ import downSVG from '@plone/volto/icons/down-key.svg';
 
 import config from '@plone/volto/registry';
 
-import {
-  getFormData,
-  exportCsvFormData,
-  clearFormData,
-  setSubblocksIDList,
-} from 'volto-form-block/actions';
+import { setSubblocksIDList } from 'volto-form-block/actions';
 
 import { BlockDataForm } from '@plone/volto/components';
-import { flattenToAppURL } from '@plone/volto/helpers/Url/Url';
 import { getFieldName } from 'volto-form-block/components/utils';
+import SidebarDataActions from 'volto-form-block/components/SidebarDataActions';
 
 import './Sidebar.css';
 
@@ -74,6 +69,7 @@ const Sidebar = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [data]);
 
+  const datatableEnabled = config.blocks.blocksConfig.form.enableDatatableView;
   return (
     <Form>
       <Segment.Group raised>
@@ -92,18 +88,27 @@ const Sidebar = ({
               });
             }}
             formData={data}
+            onChangeBlock={() => {}}
+            block={block}
           />
-          {properties?.['@components']?.form_data && (
-            <Form.Field inline>
-              <Grid></Grid>
-            </Form.Field>
+          {!datatableEnabled && (
+            <SidebarDataActions
+              properties={properties}
+              block={block}
+              data={data}
+            />
           )}
         </Segment>
         <Accordion fluid styled className="form">
           {data.subblocks &&
             data.subblocks.map((subblock, index) => {
+              const subblockData = {
+                ...subblock,
+                [`required-${subblock.field_id}`]: subblock.required,
+                [`unique-${subblock.field_id}`]: subblock.unique,
+              };
               return (
-                <div key={'subblock' + index}>
+                <div key={'subblock' + subblock.id}>
                   <Accordion.Title
                     active={selected === index}
                     index={index}
@@ -139,17 +144,26 @@ const Sidebar = ({
                     <BlockDataForm
                       schema={FieldSchema(subblock)}
                       onChangeField={(name, value) => {
+                        let key = name;
                         const update_values = {};
                         if (subblock.field_type === 'static_text') {
                           update_values.required = false;
                         }
+                        if (
+                          name === `required-${subblock.field_id}` ||
+                          name === `unique-${subblock.field_id}`
+                        ) {
+                          key = name.split('-')[0];
+                        }
                         onChangeSubBlock(index, {
                           ...subblock,
-                          [name]: value,
+                          [key]: value,
                           ...update_values,
                         });
                       }}
-                      formData={subblock}
+                      formData={subblockData}
+                      onChangeBlock={() => {}}
+                      block={block}
                     />
                   </Accordion.Content>
                 </div>